fixing some type problems and creating an UserList component for further usage
