+++ screens/FollowerListScreen.tsx
@@ -1,0 +1,9 @@
+import UserList from "../components/UserList";
+
+const FollowerListScreen = () => {
+    return (
+      <UserList />
+    );
+  }
+
+export default FollowerListScreen;
